+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - Started prototyping a flow limitation challenge
 - **Forum Changes**
   - Forum post anchors now correctly offset based on the top navigation bar
+  - Users can now post twice in a row on a forum topic
 - **Agreement Changes**
   - Step 2 now has additional instructions to clarify the importance of describing the "Specific Purpose", and selecting an appropriate number of datasets
   - The signature step now allows users to opt out of signing the agreement if they have institutional requirements forbidding signing of agreements